Initial checksum support
Checksum and tilejson are stored in a flat json file managed
by pickledb. The checksum prevents reprocessing of tiles if the
content hasn't changed.

We store this data in a flat file instead of on the resource
dictionary because there's a bug in CKAN where editing the title
of a resource will clear extra metadata keys that a resource holds.
With this commit we listen to any resource change and append the
extra metadata keys from Celery (outside CKAN).

:warning: this only works if Celery concurrency is set to 1
since access to the json flat file is blocking and sequential.