Refactor using IResourceController
* IDomainObjectModification does not support notification of resource
deletion (because it is essentially capturing dataset changes, not
resource changes). We move to implementing IResourceController and
hook into the `after_create`, `after_update`, `before_delete` and
`before_show` actions.

* Refactor of lib.py into classes for better readability and
documentation. The TileProcessor class handles the celery tasks for
create, update and delete. The CacheHandler class handles the cache
that stores the resource checksums and tilejson URLs.

* Fix a bug related that would clear the tilejson URLs from the metadata
when editing the resource through the UI. Previously, we would update
the resource metadata in celery through capturing dataset notifications.
However, updating resource metadata through celery using
IResourceController triggers an infinite loop. To avoid this, we
implement the `before_show` action, move the `tmp` directory to a volume
accessible to the web container, and read the cache before rendering
the resource.