Fix for resource URLs #59
This fixes the apparent resource URLs on the resource read and listings.
CKAN uses the docker container address to save the resource URLs in the
database. They then appear in templates with the hostname and port as
localhost:5102

We add a new function that takes the URL saved by CKAN, extracts the
path and appends it to the `ckan.site_url` in the config. This does not
modify the CKAN database, only the render methods; this fixes
downloading a resource.