Add configuration to exclude orgs from listings
This adds a configuration variable (ckanext.theme.exclude_org_list)
which can be set to a space separated list of organizations to exclude
from display in the facets of package search and the /organization page.

This is useful if an organization is used to group datasets (public or
private) but does not have members.